Remove FFI code, move contract and utils to new crate
