+++ lib/Cargo.toml
@@ -3,6 +3,7 @@ members = [
     "ain-*",
     "cli",
     "labs-grpc2",
+    "contracts",
 ]
 
 default-members = [
