+++ lib/ain-evm/src/evm.rs
@@ -1,4 +1,4 @@
-use crate::backend::{EVMBackend, EVMBackendError, Vicinity};
+use crate::backend::{EVMBackend, Vicinity};
 use crate::block::BlockService;
 use crate::core::{EVMCoreService, EVMError, NativeTxHash, MAX_GAS_PER_BLOCK};
 use crate::executor::{AinExecutor, TxResponse};
@@ -19,7 +19,6 @@ use ethereum_types::{Bloom, H160, H64, U256};
 
 use crate::bytes::Bytes;
 use crate::services::SERVICES;
-use crate::transaction::system::{DeployContractData, SystemTx};
 use anyhow::anyhow;
 use hex::FromHex;
 use log::debug;
@@ -218,20 +217,6 @@ impl EVMServices {
                         failed_transactions.push(hex::encode(hash));
                     }
                 }
-                QueueTx::SystemTx(SystemTx::DeployContract(DeployContractData {
-                    address,
-                    storage,
-                    bytecode,
-                })) => {
-                    debug!(
-                        "[finalize_block] DeployContract for address {}, storage {:#?}, bytecode {:#?}",
-                        address, storage, bytecode
-                    );
-
-                    if let Err(e) = executor.deploy_contract(address, bytecode, storage) {
-                        debug!("[finalize_block] DeployContract failed with {e}");
-                    }
-                }
             }
 
             executor.commit();
@@ -359,9 +344,7 @@ impl EVMServices {
 
     pub fn counter_contract() -> Result<(H160, Bytes, Vec<(H256, H256)>), Box<dyn Error>> {
         let address = H160::from_str("0x0000000000000000000000000000000000000301").unwrap();
-        let bytecode = get_bytecode(include_str!(
-            "../../ain-rs-exports/counter_contract/output/bytecode.json"
-        ))?;
+        let bytecode = contracts::get_counter_bytecode()?;
         let (_, latest_block_number) = SERVICES
             .evm
             .block
@@ -380,36 +363,11 @@ impl EVMServices {
 
         Ok((
             address,
-            bytecode,
-            vec![(H256::from_low_u64_be(1), u256_to_h256(count + U256::one()))],
+            Bytes::from(bytecode),
+            vec![(
+                H256::from_low_u64_be(1),
+                contracts::u256_to_h256(count + U256::one()),
+            )],
         ))
     }
-}
-
-fn u256_to_h256(input: U256) -> H256 {
-    let mut bytes = [0_u8; 32];
-    input.to_big_endian(&mut bytes);
-
-    H256::from(bytes)
-}
-
-fn get_abi_encoded_string(input: &str) -> H256 {
-    let length = input.len();
-
-    let mut storage_value = H256::default();
-    storage_value.0[31] = (length * 2) as u8;
-    storage_value.0[..length].copy_from_slice(input.as_bytes());
-
-    storage_value
-}
-
-fn get_bytecode(input: &str) -> Result<Bytes, Box<dyn Error>> {
-    let bytecode_json: serde_json::Value = serde_json::from_str(input)?;
-    let bytecode_raw = bytecode_json["object"]
-        .as_str()
-        .ok_or_else(|| anyhow!("Bytecode object not available".to_string()))?;
-
-    Ok(Bytes::from(
-        hex::decode(&bytecode_raw[2..]).map_err(|e| anyhow!(e.to_string()))?,
-    ))
 }
