+++ lib/ain-evm/src/txqueue.rs
@@ -5,7 +5,6 @@ use std::{
     sync::{Mutex, RwLock},
 };
 
-use crate::transaction::system::SystemTx;
 use crate::{
     core::NativeTxHash,
     transaction::{bridge::BridgeTx, SignedTx},
@@ -162,7 +161,6 @@ impl TransactionQueueMap {
 pub enum QueueTx {
     SignedTx(Box<SignedTx>),
     BridgeTx(BridgeTx),
-    SystemTx(SystemTx),
 }
 
 type QueueTxWithNativeHash = (QueueTx, NativeTxHash);
@@ -232,7 +230,6 @@ impl TransactionQueue {
             let tx_sender = match tx {
                 QueueTx::SignedTx(tx) => tx.sender,
                 QueueTx::BridgeTx(tx) => tx.sender(),
-                QueueTx::SystemTx(_) => H160::zero(),
             };
             tx_sender != sender
         });
