+++ lib/ain-rs-exports/Cargo.toml
@@ -27,7 +27,3 @@ cxx = "1.0"
 [build-dependencies]
 cxx-gen = "0.7"
 proc-macro2 = "1.0"
-ethers = "2.0.7"
-ethers-solc = "2.0.7"
-serde_json = "1.0"
-anyhow = "1.0"
