+++ lib/ain-rs-exports/build.rs
@@ -1,11 +1,9 @@
 use proc_macro2::TokenStream;
 
-use anyhow::anyhow;
-use ethers_solc::{Project, ProjectPathsConfig};
+use std::env;
 use std::fs::File;
 use std::io::{Read, Write};
 use std::path::PathBuf;
-use std::{env, fs};
 
 fn main() -> Result<(), Box<dyn std::error::Error>> {
     let pkg_name = env::var("CARGO_PKG_NAME")?;
@@ -56,45 +54,6 @@ fn main() -> Result<(), Box<dyn std::error::Error>> {
             git_head_path.to_str().ok_or("git head path err")?
         );
     }
-
-    // compile solidity project
-    // configure `root` as our project root
-    let root = PathBuf::from("counter_contract");
-    if !root.exists() {
-        return Err("Project root {root:?} does not exists!".into());
-    }
-
-    let paths = ProjectPathsConfig::builder()
-        .root(&root)
-        .sources(&root)
-        .build()?;
-
-    let project = Project::builder()
-        .paths(paths)
-        .set_auto_detect(true)
-        .no_artifacts()
-        .build()?;
-    let output = project.compile().unwrap();
-    let artifacts = output.into_artifacts();
-
-    for (id, artifact) in artifacts {
-        if id.name == "Counter" {
-            let abi = artifact.abi.ok_or_else(|| anyhow!("ABI not found"))?;
-            let bytecode = artifact.deployed_bytecode.expect("No bytecode found");
-
-            fs::create_dir_all("counter_contract/output/")?;
-            fs::write(
-                PathBuf::from("counter_contract/output/bytecode.json"),
-                serde_json::to_string(&bytecode).unwrap().as_bytes(),
-            )?;
-            fs::write(
-                PathBuf::from("counter_contract/output/abi.json"),
-                serde_json::to_string(&abi).unwrap().as_bytes(),
-            )?;
-        }
-    }
-
-    project.rerun_if_sources_changed();
 
     Ok(())
 }
